+++ server/src/entities/Service.ts
@@ -1,4 +1,6 @@
+import { Field, ObjectType } from "type-graphql";
 import {
+  BaseEntity,
   CreateDateColumn,
   Entity,
   ManyToMany,
@@ -6,23 +8,33 @@ import {
   PrimaryColumn,
   UpdateDateColumn,
 } from "typeorm";
-import { User } from "./User";
 import { Document } from "./Document";
+import { User } from "./User";
 
+@ObjectType()
 @Entity()
-export class Service {
+export class Service extends BaseEntity {
+  @Field()
   @PrimaryColumn()
   name: string;
 
+  @Field()
   @CreateDateColumn()
   createdAt: Date;
 
+  @Field()
   @UpdateDateColumn()
   updatedAt: Date;
 
-  @OneToMany(() => User, (user) => user.service)
+  @OneToMany(() => User, (user) => user.service, {
+    nullable: true,
+    cascade: false,
+  })
   user: User[];
 
-  @ManyToMany(() => Document, document => document.services)
+  @ManyToMany(() => Document, (document) => document.services, {
+    nullable: true,
+    cascade: false,
+  })
   documents: Document[];
 }